+++ app/Http/Controllers/ConcursoController.php
@@ -5,7 +5,6 @@ namespace App\Http\Controllers;
 use App\Events\CrudEvent;
 use App\Traits\TabelaAdmin;
 use Illuminate\Http\Request;
-use App\Traits\ControleAcesso;
 use App\Http\Requests\ConcursoRequest;
 use App\Http\Controllers\CrudController;
 use App\Repositories\ConcursoRepository;
@@ -13,7 +12,7 @@ use Illuminate\Support\Facades\Request as IlluminateRequest;
 
 class ConcursoController extends Controller
 {
-    use ControleAcesso, TabelaAdmin;
+    use TabelaAdmin;
 
     private $class = 'ConcursoController';
     private $concursoRepository;
@@ -39,12 +38,12 @@ class ConcursoController extends Controller
 
     public function index()
     {
-        $this->autoriza($this->class, __FUNCTION__);
+        $this->authorize('viewAny', auth()->user());
 
         $resultados = $this->concursoRepository->getToTable();
         $tabela = $this->tabelaCompleta($resultados);
 
-        if(!$this->mostra($this->class, 'create')) {
+        if(auth()->user()->cannot('create', auth()->user())) {
             unset($this->variaveis['btn_criar']);
         }        
         $variaveis = (object) $this->variaveis;
@@ -54,7 +53,7 @@ class ConcursoController extends Controller
 
     public function create()
     {
-        $this->autoriza($this->class, __FUNCTION__);
+        $this->authorize('create', auth()->user());
 
         $variaveis = (object) $this->variaveis;
 
@@ -63,6 +62,8 @@ class ConcursoController extends Controller
 
     public function store(ConcursoRequest $request)
     {
+        $this->authorize('create', auth()->user());
+
         $request->validated();
 
         $save = $this->concursoRepository->store($request);
@@ -80,7 +81,7 @@ class ConcursoController extends Controller
 
     public function edit($id)
     {
-        $this->autoriza($this->class, __FUNCTION__);
+        $this->authorize('updateOther', auth()->user());
 
         $resultado = $this->concursoRepository->getById($id);
         $variaveis = (object) $this->variaveis;
@@ -90,6 +91,8 @@ class ConcursoController extends Controller
 
     public function update(ConcursoRequest $request, $id)
     {
+        $this->authorize('updateOther', auth()->user());
+
         $request->validated();
         
         $update = $this->concursoRepository->update($id, $request); 
@@ -116,7 +119,7 @@ class ConcursoController extends Controller
 
     public function destroy($id)
     {
-        $this->autoriza($this->class, __FUNCTION__);
+        $this->authorize('delete', auth()->user());
 
         $delete = $this->concursoRepository->destroy($id);
 
@@ -133,7 +136,7 @@ class ConcursoController extends Controller
 
     public function lixeira()
     {
-        $this->autorizaStatic(['1']);
+        $this->authorize('onlyAdmin', auth()->user());
 
         $resultados = $this->concursoRepository->getTrashed();
         $tabela = $this->tabelaTrashed($resultados);
@@ -144,7 +147,7 @@ class ConcursoController extends Controller
 
     public function restore($id)
     {
-        $this->autorizaStatic(['1']);
+        $this->authorize('onlyAdmin', auth()->user());
 
         $restore = $this->concursoRepository->restore($id);
 
@@ -228,11 +231,11 @@ class ConcursoController extends Controller
         $contents = $query->map(function($row) {
             $acoes = '<a href="'.route('concursos.show', $row->idconcurso).'" class="btn btn-sm btn-default">Ver</a> ';
             
-            if($this->mostra('ConcursoController', 'edit')) {
+            if(auth()->user()->can('updateOther', auth()->user())) {
                 $acoes .= '<a href="'.route('concursos.edit', $row->idconcurso).'" class="btn btn-sm btn-primary">Editar</a> ';
             }
                 
-            if($this->mostra('ConcursoController', 'destroy')) {
+            if(auth()->user()->can('delete', auth()->user())) {
                 $acoes .= '<form method="POST" action="'.route('concursos.destroy', $row->idconcurso).'" class="d-inline">';
                 $acoes .= '<input type="hidden" name="_token" value="'.csrf_token().'" />';
                 $acoes .= '<input type="hidden" name="_method" value="delete" />';
